update the User profile pages
create the UserService interface and implements the UserServiceImplement class inside
